Webinar fixes
